refactor: expansion_heredoc.cのファイル分割, connect_expansion_list_heredoc
